Add files via upload
In my code, I defined the whole thing as a function. First, I check for the cases where there is less than 2 days/stock prices or if the max price is after the min price. Then if the scenario is very complicated, I use a nested for loop to go over the whole array to compare from right number to left number and by subtracting the right by left numbers. From this we compare the values and find the max value possible from all the possible profit values and return the highest value.